Preliminary conversion of Garnet Manual to Texinfo
The old Garnet manual was written in a Texinfo-like markup languange,
Scribe. Converting the manual to Texinfo involves editing the old
sounce files which have a *.mss extention and refactoring the Scribe
markup into Texinfo markup. In many cases that's trivial,
e.g. @b{bolditem} instead of @B(bolditem). In other cases it involves
more effort.

The current state of the conversion is that enough of the refactoring
has been done to build the new manual into a PDF file. This involved
the following tasks:

- Converting all included graphics files to PDF. Ultimately they
  should be converted to *.png files so they can be properly
  clipped. only a few files have been converted to PNGs so far.

- All the previous Scribe text has been consolidated into two files:
  kr-manual.texi and overview.texi. The reason for this is that much
  of the conversion can be done by text substitution. Having all the
  text in one or two files makes this easier to do. Once the
  refactoring is complete, the text should probably be broken up again
  into various sections to make editing easier.

- Some tweeks have had to be made to the build process the machinery
  of which is mostly contained in doc/src/makefile and
  doc/src/garnet-manual.texi

The remaining work is mostly editorial and tweaking the grophics. Of
course the manual itself is somewhat out of date. For example it
describes how to build Garnet using the old Garnet loader, while the
current system can only be loaded using ASDF.